重置消息页面

+++ src/Models/Models.BiliBili/User/AtMessageResponse.cs
@@ -46,7 +46,7 @@ namespace Bili.Models.BiliBili
         /// @我的时间.
         /// </summary>
         [JsonProperty(NullValueHandling = NullValueHandling.Ignore, PropertyName = "at_time", Required = Required.Default)]
-        public int AtTime { get; set; }
+        public long AtTime { get; set; }
     }
 
     /// <summary>
